Backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
asgiref==3.7.2
click==8.1.6
colorama==0.4.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
fastapi==0.100.1
h11==0.14.0
httptools==0.6.0
idna==3.4
joblib==1.3.1
numpy==1.25.1
pandas==2.0.3
pydantic==2.1.1
pydantic_core==2.4.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.11.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
threadpoolctl==3.2.0
typing_extensions==4.7.1
tzdata==2023.3
uvicorn==0.23.1
watchfiles==0.19.0
websockets==11.0.3
xgboost==1.7.6
